+++ app/code/community/Lyranetwork/Payzen/Model/Api/Ws/WsResponse.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * PayZen V2-Payment Module version 1.10.3 for Magento 1.4-1.9. Support contact : support@payzen.eu.
+ * PayZen V2-Payment Module version 1.11.0 for Magento 1.4-1.9. Support contact : support@payzen.eu.
  *
  * @category  Payment
  * @package   Payzen